backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy==2.0.40
psycopg[binary]==3.2.6
pydantic-settings==2.8.1
python-jose[cryptography]==3.4.0
bcrypt==4.0.1
passlib==1.7.4
email-validator==2.2.0
